Quick wins: rename kebab-case files, inline single-use export, remove re-exports
- Rename persisted-flow.ts → persistedFlow.ts, round-persisted-flow.ts → roundPersistedFlow.ts
- Delete round-bounds.ts (inline trivial isRoundAtOrBeyondLimit check)
- Remove SubagentProgressUpdate re-exports from @tools/subagentResults; consumers import from @shared/schemas directly
- Update all consumer imports to match new file names
