+++ src/tools/ExecutionsTool.ts
@@ -26,7 +26,7 @@ import {
   type TodoEntry,
   listExecutions,
 } from '@agent/storage';
-import { flowKey } from '@agent/node/persisted-flow';
+import { flowKey } from '@agent/node/persistedFlow';
 import { getCurrentToolFileInteractionContext } from '@agent/toolUse/ToolFileInteractionContext';
 import {
   type ExecutionHandle,
